+++ pyAitu/__init__.py
@@ -1,11 +1,13 @@
 from .bot import Bot
 from .dispatcher import Dispatcher
-from . import types
+from . import commands
+from . import models
 from .utils import executor
 
 __all__ = [
     'Bot',
     'Dispatcher',
-    'types',
-    'executor'
+    'commands',
+    'executor',
+    'models'
 ]
