+++ pyAitu/models/__init__.py
@@ -1,0 +1,14 @@
+from .message import Message, ContentType
+from .update import Update
+from .quick_button_selected import QuickButtonSelected
+from .inline_command_selected import InlineCommandSelected
+from .media import Media
+
+__all__ = [
+    Message,
+    ContentType,
+    Update,
+    QuickButtonSelected,
+    InlineCommandSelected,
+    Media
+]
