+++ pyAitu/models/image.py
