+++ pyAitu/models/recipient.py
@@ -1,4 +1,4 @@
-from ..utils.strings import DEFAULT_USER, TYPE, ID
+from pyAitu.utils.strings import DEFAULT_USER, TYPE, ID
 
 
 class Recipient:
