--- pyAitu/types/__init__.py
@@ -1,22 +1,0 @@
-from .message import Message, ContentType
-from .update import Update
-from .command import Command
-from .quick_button_command import QuickButtonCommand
-from .quick_button_selected import QuickButtonSelected
-from .inline_command import InlineCommand
-from .inline_command_selected import InlineCommandSelected
-from .reply_command import ReplyCommand
-from .media import Media
-
-__all__ = [
-    'Message',
-    'Update',
-    'Command',
-    'QuickButtonCommand',
-    'QuickButtonSelected',
-    'InlineCommand',
-    'InlineCommandSelected',
-    'ReplyCommand',
-    'Media',
-    'ContentType'
-]
